Move predict integration tests to fixtures
