+++ test-integration/test_integration/fixtures/string-project/predict.py
@@ -1,0 +1,7 @@
+import cog
+
+
+class Predictor(cog.Predictor):
+    @cog.input("input", type=str)
+    def predict(self, input):
+        return "hello " + input
