+++ test-integration/test_integration/fixtures/subdirectory-project/cog.yaml
@@ -1,0 +1,3 @@
+build:
+  python_version: "3.8"
+predict: "my-subdir/predict.py:Predictor"
