+++ test-integration/test_integration/fixtures/subdirectory-project/my-subdir/predict.py
@@ -1,0 +1,9 @@
+import cog
+
+from mylib import concat
+
+
+class Predictor(cog.Predictor):
+    @cog.input("input", type=str)
+    def predict(self, input):
+        return concat("hello", input)
