+++ test-integration/test_integration/fixtures/subdirectory-project/mylib.py
@@ -1,0 +1,2 @@
+def concat(a, b):
+    return a + " " + b
